Rename ResizableVisualizer to ResizeTooltip
Also adjust properties with axis and isEnabled properties
