+++ packages/components/src/resize-tooltip/styles/resize-tooltip.styles.js
